pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bybit-mcp"
version = "0.1.0"
description = "Bybit Trading MCP Server for Claude - Cloud Run deployable"
requires-python = ">=3.12"
dependencies = [
    "mcp[cli]>=1.12",
    "pybit>=5.8",
    "uvicorn>=0.34",
    "python-dotenv>=1.0",
    "starlette>=0.45",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
